watcher: adapt logger and cleanup fixes #15
